added app name to about
