+++ ecommerce-order-service-api/src/main/resources/application-local.yml
@@ -2,6 +2,7 @@ buildNumber: unknown
 buildTime: unknown
 gitRevision: unknown
 gitBranch: unknown
+appName: unknown
 
 spring:
   datasource:
