+++ ecommerce-order-service-sdk/src/main/java/com/ecommerce/order/sdk/representation/about/AboutRepresentation.java
@@ -1,0 +1,16 @@
+package com.ecommerce.order.sdk.representation.about;
+
+
+import lombok.Value;
+
+@Value
+public class AboutRepresentation {
+    private String appName;
+    private String buildNumber;
+    private String buildTime;
+    private String deployTime;
+    private String gitRevision;
+    private String gitBranch;
+    private String environment;
+
+}
